refactor: 패키지 트리 변경
기존: 웹 패키지 콘솔 패키지 분리
변경: 패키지 통합
