+++ src/main/java/racingcar/ConsoleApplication.java
@@ -1,0 +1,10 @@
+package racingcar;
+
+import racingcar.controller.ConsoleController;
+
+public class ConsoleApplication {
+    public static void main(String[] args) {
+        ConsoleController consoleController = new ConsoleController();
+        consoleController.runGame();
+    }
+}
