+++ src/main/java/racingcar/WebApplication.java
@@ -1,0 +1,10 @@
+package racingcar;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebApplication { public static void main(String[] args) {
+        SpringApplication.run(WebApplication.class, args);
+    }
+}
