+++ src/main/java/racingcar/util/NumberGenerator.java
@@ -1,0 +1,5 @@
+package racingcar.util;
+
+public interface NumberGenerator {
+    int generateNumber();
+}
