--- src/main/java/racingcar/web/util/NumberGenerator.java
@@ -1,5 +1,0 @@
-package racingcar.web.util;
-
-public interface NumberGenerator {
-    int generateNumber();
-}
